issue: 3777348 Merging socket_fd_api with sockinfo
Removing redundant base from sockinfo allows to put data path members near the vptr for better cache utilization.
